app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Teemu Keskitalo <teemu.keskitalo@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.31.0"
fastapi = "^0.110.1"
pydantic = "^2.7.0"
uvicorn = "^0.29.0"
todoist-api-python = "^2.1.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
